Lazy evaluation.  eval doesn't evaluate it's arguments, it just creates a Thunk with the form and bindings it was supposed to evaluate.  The thunk is evaluated only if/when the value is actually needed.
Except eval does atom.eval(binding) because it's cheap and why not.

Note that this is the correct place to create thunks.  The paper "Cons
should not evaluate its arguments" is full of crap.  It should be
"Eval should not evaluate its arguments."  Thunking in cons and unthunking
on car/cdr leads to infinte loops when filtering an infinite list where
nothing matches the predicate because it never conses anything.